# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lab Resource Utilization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /utilization:
    get:
      operationId: listUtilizationRecords
      tags: [utilization]
      summary: List utilization records
      parameters:
        - name: equipmentId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Utilization records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UtilizationRecord"
  /utilization/stats:
    get:
      operationId: getEquipmentUtilizationStats
      tags: [utilization]
      summary: Get utilization stats for equipment
      parameters:
        - name: equipmentId
          in: query
          required: true
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Utilization stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtilizationStats"
  /utilization/heatmap:
    get:
      operationId: getUtilizationHeatmap
      tags: [utilization]
      summary: Get utilization heatmap data
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: institutionId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapDataPoint"
...
  /maintenance:
    get:
      operationId: listMaintenanceRecords
      tags: [maintenance]
      summary: List maintenance records
      parameters:
        - name: equipmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRecord"
    post:
      operationId: createMaintenanceRecord
      tags: [maintenance]
      summary: Create maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Created maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
  /maintenance/{maintenanceId}:
    get:
      operationId: getMaintenanceRecord
      tags: [maintenance]
      summary: Get maintenance record by ID
      parameters:
        - name: maintenanceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
    patch:
      operationId: updateMaintenanceRecord
      tags: [maintenance]
      summary: Update maintenance record
      parameters:
        - name: maintenanceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceUpdate"
      responses:
        "200":
          description: Updated maintenance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
  /maintenance/upcoming:
    get:
      operationId: getUpcomingMaintenance
      tags: [maintenance]
      summary: Get upcoming maintenance tasks
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Upcoming maintenance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRecord"

  # ─── INTER-INSTITUTION SHARING ────────────────────────────────────────────
  /sharing/agreements:
    get:
      operationId: listSharingAgreements
      tags: [sharing]
      summary: List sharing agreements
      responses:
        "200":
          description: List of sharing agreements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharingAgreement"
    post:
      operationId: createSharingAgreement
      tags: [sharing]
      summary: Create sharing agreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharingAgreementInput"
      responses:
        "201":
          description: Created agreement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharingAgreement"
  /sharing/agreements/{agreementId}:
    get:
      operationId: getSharingAgreement
      tags: [sharing]
      summary: Get sharing agreement by ID
      parameters:
        - name: agreementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sharing agreement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharingAgreement"
    patch:
      operationId: updateSharingAgreement
      tags: [sharing]
      summary: Update sharing agreement status
      parameters:
        - name: agreementId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharingAgreementUpdate"
      responses:
        "200":
          description: Updated agreement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharingAgreement"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    UtilizationRecord:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        bookingId:
          type: ["integer", "null"]
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        utilizationPercent:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - equipmentId
        - startTime
        - createdAt
    UtilizationStats:
      type: object
      properties:
        equipmentId:
          type: integer
        equipmentName:
          type: string
        totalHoursAvailable:
          type: number
        totalHoursUsed:
          type: number
        utilizationRate:
          type: number
        peakHours:
          type: string
        idleHours:
          type: number
        bookingCount:
          type: integer
      required:
        - equipmentId
        - equipmentName
        - totalHoursAvailable
        - totalHoursUsed
        - utilizationRate
        - bookingCount
    HeatmapDataPoint:
      type: object
      properties:
        hour:
          type: integer
        dayOfWeek:
          type: integer
        utilizationCount:
          type: integer
        utilizationRate:
          type: number
      required:
        - hour
        - dayOfWeek
        - utilizationCount
        - utilizationRate
...
    MaintenanceRecord:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        type:
          type: string
          enum: [PREVENTIVE, CORRECTIVE, CALIBRATION, INSPECTION]
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        scheduledDate:
          type: string
        completedDate:
          type: ["string", "null"]
        assignedTechnicianId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        downtimeHours:
          type: ["number", "null"]
        createdAt:
          type: string
        equipmentName:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
      required:
        - id
        - equipmentId
        - type
        - status
        - scheduledDate
        - createdAt
    MaintenanceInput:
      type: object
      properties:
        equipmentId:
          type: integer
        type:
          type: string
          enum: [PREVENTIVE, CORRECTIVE, CALIBRATION, INSPECTION]
        scheduledDate:
          type: string
        assignedTechnicianId:
          type: integer
        description:
          type: string
        cost:
          type: number
        notes:
          type: string
      required:
        - equipmentId
        - type
        - scheduledDate
    MaintenanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        completedDate:
          type: string
        assignedTechnicianId:
          type: integer
        notes:
          type: string
        cost:
          type: number
        downtimeHours:
          type: number
    SharingAgreement:
      type: object
      properties:
        id:
          type: integer
        requestingInstitutionId:
          type: integer
        owningInstitutionId:
          type: integer
        equipmentId:
          type: integer
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, EXPIRED]
        dailyRate:
          type: ["number", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        terms:
          type: ["string", "null"]
        createdAt:
          type: string
        equipmentName:
          type: ["string", "null"]
        requestingInstitutionName:
          type: ["string", "null"]
        owningInstitutionName:
          type: ["string", "null"]
      required:
        - id
        - requestingInstitutionId
        - owningInstitutionId
        - equipmentId
        - status
        - createdAt
    SharingAgreementInput:
      type: object
      properties:
        requestingInstitutionId:
          type: integer
        owningInstitutionId:
          type: integer
        equipmentId:
          type: integer
        dailyRate:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        terms:
          type: string
      required:
        - requestingInstitutionId
        - owningInstitutionId
        - equipmentId
    SharingAgreementUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, EXPIRED]
        terms:
          type: string
      required:
        - status
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedEntityType:
          type: ["string", "null"]
        relatedEntityId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt
...