# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lab Resource Utilization Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceInputType } from './maintenanceInputType';

export interface MaintenanceInput {
  equipmentId: number;
  type: MaintenanceInputType;
  scheduledDate: string;
  assignedTechnicianId?: number;
  description?: string;
  cost?: number;
  notes?: string;
}